requirements.txt
sqlalchemy==2.0.40
pymysql==1.1.2
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.1.0
pytest==8.3.5
snowflake-connector-python[pandas]==3.14.0
